requirements.txt
discord.py
twitchio
google-generativeai
python-dotenv